pyproject.toml
[tool.poetry]
name = "tapir"
version = "0.1.0"
description = ""
authors = ["Michael Kleimann <dev@mkleimann.com>", "Théophile Madet <theo.madet@posteo.net>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
Django = "^5.2"
django-extensions = "^3.2.3"
django-bootstrap-datepicker-plus = "^5.0.2"
psycopg2-binary = "^2.9.10"
django-tables2 = "^2.7.5"
django-filter = "^25.1"
tablib = "^3.8.0"
django-phonenumber-field = { extras = ["phonenumbers"], version = "^5.2.0" }
phonenumbers = "^8.12.44"
weasyprint = "^65.0"
django-select2 = "^8.4.0"
django-silk = "^5.3.2"
gunicorn = "^23.0.0"
django-environ = "^0.12.0"
django-bootstrap5 = "^25.1"
django-formtools = "^2.5.1"
celery = "^5.5.0"
redis = "^5.2.1"
django-localflavor = "^4.0"
tomlkit = "^0.13.2"
nanoid = "^2.0.0"
lxml= "^5.3.2"
python-keycloak = "^5.3.1"
PyJWT = "^2.10.1"
unidecode = "^1.3.8"
importlib-resources = "^6.5.2"
tapir-mail = { url = "https://github.com/FoodCoopX/tapir-mail-releases/releases/download/0.0.0-dev-1747395568/tapir_mail-0.0.0.dev1747395568-py3-none-any.whl" }
djangorestframework = "^3.16.0"
drf-spectacular = "^0.28.0"
django-vite = "^3.1.0"
django-weasyprint = "^2.4.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.17.0"
Werkzeug = "^3.1.3"
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
pytest-django = "^4.11.1"
pytest-sugar = "^1.0.0"
selenium = "^4.31.0"
factory-boy = "^3.3.3"
black = "^25.1.0"
ptvsd = "^4.3.2"
icecream = "^2.1.4"
djlint = "^1.36.4"
ipython = "^9.0.2"
django-upgrade = "^1.24.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.djlint]
profile="django"

[tool.coverage.run]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/urls.py",
    "migrate.py"
]
relative_files = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tapir.settings"
python_files = ["tests.py", "test_*.py", "tests_*.py"]
testpaths = "tapir"
